% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvshear.R
\name{BVshear}
\alias{BVshear}
\alias{dBVshear}
\alias{pBVshear}
\alias{qBVshear}
\alias{rBVshear}
\title{Brutsaert-Veron shear driven distribution}
\usage{
dBVshear(x, u, ...)

pBVshear(q, u, ...)

qBVshear(p, u, ...)

rBVshear(n, u, ...)
}
\arguments{
\item{x}{vector of quantiles.}

\item{u}{friction velocity.}

\item{...}{additional arguments to be passed to the Pearson Type VI distribution of the PearsonDS package.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations for random generation.}
}
\value{
\code{dBVshear} gives the density, \code{pBVshear} gives the distribution function, \code{qBVshear} gives the quantile function, and \code{rBVshear} provides the random generation.
}
\description{
Density, distribution function, quantile function and random generation for the Brutsaert-Veron shear driven distribution.
}
\note{
These functions are just wrappers of the Pearson Type VI distribution functions of the package \code{PearsonDS} applied to the context of the surface renewal distribution. See reference below.
}
\examples{
# density at 0.03 for u = 0.5
dBVshear(x = 0.03, u = 0.5) 

# log-density at 0.03 for u = 0.5
dBVshear(x = 0.03, u = 0.5, log = TRUE) 

# P(X <= 0.03) for u = 0.5
pBVshear(q = 0.03, u = 0.5) 

# P(X > 0.03) for u = 0.5
pBVshear(q = 0.03, u = 0.5, lower.tail = FALSE) 

# median of the distribution for u = 0.5
qBVshear(p = 0.5, u = 0.5)

set.seed(1234)
tau <- rBVshear(n = 1E4, u = 0.5)
eps <- 0.005
hist(tau, freq = FALSE, breaks = seq(0, max(tau) + eps, eps), xlim = c(0, 0.2), 
     ylim = c(0, 40), xlab = "Contact time", main = "Histogram of contact times")
}
\references{
Solheid, B.S., Dias, N.L.C., and Costa, E.G. (2023). The equations for the nondimensional moments of the surface renewal distribution.
}
